Switch implementation of Python bindings to C++
Allows C++ to be used when creating bindings for Python.
