+++ python/hawkey/CMakeLists.txt
@@ -24,26 +24,26 @@ SET (hawkey_COPIES __init__.py dbg.py)
 FILE(COPY ${hawkey_COPIES} DESTINATION ./hawkey/)
 
 SET (hawkeymodule_SRCS
-    advisory-py.c
-    advisorypkg-py.c
-    advisoryref-py.c
-    exception-py.c
-    goal-py.c
-    hawkeymodule.c
-    iutil-py.c
-    nevra-py.c
-    module-form-py.c
-    package-py.c
-    packagedelta-py.c
-    possibilities-py.c
-    query-py.c
-    repo-py.c
-    reldep-py.c
-    sack-py.c
-    selector-py.c
-    solution-py.c
-    subject-py.c
-    pycomp.c)
+    advisory-py.cpp
+    advisorypkg-py.cpp
+    advisoryref-py.cpp
+    exception-py.cpp
+    goal-py.cpp
+    hawkeymodule.cpp
+    iutil-py.cpp
+    module-form-py.cpp
+    nevra-py.cpp
+    package-py.cpp
+    packagedelta-py.cpp
+    possibilities-py.cpp
+    query-py.cpp
+    reldep-py.cpp
+    repo-py.cpp
+    sack-py.cpp
+    selector-py.cpp
+    solution-py.cpp
+    subject-py.cpp
+    pycomp.cpp)
 
 ADD_LIBRARY(_hawkeymodule SHARED ${hawkeymodule_SRCS})
 SET_TARGET_PROPERTIES(_hawkeymodule PROPERTIES PREFIX "")
