+++ python/hawkey/advisory-py.hpp
