+++ python/hawkey/advisorypkg-py.hpp
