+++ python/hawkey/advisoryref-py.hpp
