+++ python/hawkey/exception-py.hpp
