+++ python/hawkey/goal-py.hpp
