+++ python/hawkey/hawkeymodule.cpp
@@ -36,25 +36,25 @@
 #include "dnf-version.h"
 
 // pyhawkey
-#include "advisory-py.h"
-#include "advisorypkg-py.h"
-#include "advisoryref-py.h"
-#include "exception-py.h"
-#include "goal-py.h"
-#include "nevra-py.h"
-#include "module-form-py.h"
-#include "package-py.h"
-#include "packagedelta-py.h"
-#include "possibilities-py.h"
-#include "query-py.h"
-#include "reldep-py.h"
-#include "repo-py.h"
-#include "sack-py.h"
-#include "selector-py.h"
-#include "solution-py.h"
-#include "subject-py.h"
+#include "advisory-py.hpp"
+#include "advisorypkg-py.hpp"
+#include "advisoryref-py.hpp"
+#include "exception-py.hpp"
+#include "goal-py.hpp"
+#include "nevra-py.hpp"
+#include "module-form-py.hpp"
+#include "package-py.hpp"
+#include "packagedelta-py.hpp"
+#include "possibilities-py.hpp"
+#include "query-py.hpp"
+#include "reldep-py.hpp"
+#include "repo-py.hpp"
+#include "sack-py.hpp"
+#include "selector-py.hpp"
+#include "solution-py.hpp"
+#include "subject-py.hpp"
 
-#include "pycomp.h"
+#include "pycomp.hpp"
 
 static PyObject *
 detect_arch(PyObject *unused, PyObject *args)
