+++ python/hawkey/packagedelta-py.hpp
