+++ python/hawkey/query-py.hpp
