+++ python/hawkey/solution-py.hpp
